+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Confirm dialogs now abort on escape
 - The shortcut for opening settings now works as intented
 - Settings no longer collide with other modals
+- An issue where children weren't removed when their parent group was removed
 ### Changed
 - Keyboard shortcuts now use hotkeys-js for better stability
 - The item.apply event has been changed to item.change